requirements/local.txt
asgiref==3.2.10
backports.zoneinfo==0.2.1
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
django==3.1.14
django-allauth==0.42.0
django-cors-headers==3.4.0
django-debug-toolbar==3.2.1
django-dirtyfields==1.7.0
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-gdpr-assist==1.3.0
django-import-export==2.5.0
django-mjml==0.11.0
django-model-utils==4.0.0
django-money==2.1.1
django-nested-admin==3.3.3
django-ordered-model==3.4.3
django-rest-auth[with_social]==0.9.5
django-storages==1.9.1
django-yaa-settings==1.1.0
djangorestframework==3.11.2
djangorestframework-api-key==2.2.0
drf-yasg==1.17.1
et-xmlfile==1.1.0
idna==2.10
inflection==0.5.0
isort==5.10.1
itypes==1.2.0
jinja2==2.11.3
markuppy==1.14
markupsafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==20.4
psycopg2==2.8.3
pycountry==22.3.5
pycparser==2.20
pyjwt==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.2
python-magic==0.4.24
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2020.1
pyyaml==6.0
psycopg2-binary==2.9.3
requests==2.24.0
requests-oauthlib==1.3.0
ruamel-yaml==0.16.10
ruamel-yaml-clib==0.2.6
six==1.15.0
sqlparse==0.3.1
tablib[html,ods,xls,xlsx,yaml]==3.1.0
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0